clean: Moved helper methods out of main file for easier readability
